Add tests for I/O routines
In preparation for fixing #46, add a test suite that exercises the
functionality (test_save_pha). It is marked as failing.

Also add in some other basic I/O changes. Note that ideally some of
this would be in a test suite for sherpa.astro.io, but it is
currently hard to do so, since the module expects either the crates
or astropy backends to be available (i.e. it does not handle the
case of neither being available). The tests are therefore added to
test_astro_ui.py. Note that the other tests added here were written
so as to exercise parts of the code base that were not covered by
existing tests.