Move test code out from sherpa.utils
Release Notes

There is no functional change in the PR.

Move the test code out from sherpa.utils to sherpa.utils.test and
update the test code to match. Some minor PEP8 cleanup (mainly handling
of space characters, but several changes to use explicit import
statements).

Notes

This is based on the discussion of PR #141 where it was noted that
the sherpa.utils module is too large and needs separating out. This
is the first step in the process.

Test

The test suite passes with these changes, with the same number of tests
run (and tests skipped, as the tests were ran without ds9 or XSPEC
support) as before the code change.